requirements.txt
asgiref==3.7.2
attrs==23.1.0
certifi==2023.7.22
charset-normalizer==3.3.2
colorama==0.4.6
Django==4.2.7
django-filter==23.4
djangorestframework==3.14.0
drf-spectacular==0.26.5
flake8==6.1.0
gunicorn==21.2.0
idna==3.4
inflection==0.5.1
iniconfig==2.0.0
jsonschema==4.20.0
jsonschema-specifications==2023.11.1
mccabe==0.7.0
packaging==23.2
Pillow==10.1.0
pluggy==1.3.0
psycopg==3.1.14
psycopg-binary==3.1.14
py4j==0.10.9.7
pycodestyle==2.11.1
pyflakes==3.1.0
pytest==7.4.3
pytest-django==4.7.0
pytz==2023.3.post1
PyYAML==6.0.1
referencing==0.31.0
requests==2.31.0
rpds-py==0.13.0
sqlparse==0.4.4
typing_extensions==4.8.0
tzdata==2023.3
uritemplate==4.1.1
urllib3==2.1.0
